move importing functionality to appropriate place
